ai-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
httpx==0.27.0
python-dotenv==1.0.1
langchain==0.1.14
langchain-community==0.0.31
langchain-google-genai==1.0.1
faiss-cpu==1.8.0
sentence-transformers==2.6.1
pypdf==4.1.0
python-multipart==0.0.9
